src/main/java/com/o0u0o/sell/service/impl/ProductInfoServiceImpl.java
package com.o0u0o.sell.service.impl;

import com.o0u0o.sell.dataobject.ProductInfo;
import com.o0u0o.sell.enums.ProductStatus;
import com.o0u0o.sell.repository.ProductInfoRepository;
import com.o0u0o.sell.service.ProductInfoService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @Author aiuiot
 * @Date 2020/2/4 6:22 下午
 * @Descripton:
 **/
@Service
public class ProductInfoServiceImpl implements ProductInfoService {

    @Autowired
    private ProductInfoRepository repository;

    @Override
    public ProductInfo findOne(String productId) {
        return repository.findOne(productId);
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return repository.findByProductStatus(ProductStatus.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return repository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return repository.save(productInfo);
    }
}
